changes
fixed controllers
