+++ routes/api/notices.js
@@ -1,20 +1,16 @@
 const express = require("express");
 const router = express.Router();
 
-const { authentificate } = require("../../middlewares/authentificate");
-
-const { validateBody } = require("../../helpers/validatebody");
-
-const { schemas } = require("../../models/pets");
-
 const {
   getController,
   getByIdController,
   getByCategoryController,
-} = require("../../controllers/petsControllers");
+  getBySearchController,
+} = require("../../controllers/noticesControllers");
 
 router.get("/", getController);
 router.get("/notice/:noticeId", getByIdController);
+router.get("/:category", getBySearchController);
 router.get("/category/:category", getByCategoryController);
 
 module.exports = router;
